Include header strings.h
strncasecmp is a non-standard method and does not exist in the
string.h. It is defined in POSIX standard and BSD standard. The header
strings.h explicitly has the function.

POSIX and BSD compliant systems, fills in strings.h functions if
string.h is included. For a system which does not support the
non-standard functions the build will fail which is better than relying
on shimming.